Move Alias and WeakAlias into alias module. Return None in __setattr__. Raise AttributeException in WeakAlias.__delattr__ for missing key.
