+++ python/tlm_adjoint/base_equations.py
@@ -567,47 +567,6 @@ class Equation:
         raise EquationException("Method not overridden")
 
 
-class WeakAlias:
-    """
-    Alias of obj, holding no reference to obj, and valid after deallocation of
-    obj. Intended to be used in combination with weakref.finalize(obj, ...).
-    """
-
-    def __init__(self, obj):
-        if hasattr(obj, "__slots__"):
-            # Weak references to obj not possible, has attributes not
-            # accessible via __dict__ attribute
-            raise EquationException("Cannot alias object with __slots__ "
-                                    "attribute")
-        if isinstance(obj, WeakAlias):
-            raise EquationException("Cannot alias WeakAlias")
-        super().__setattr__("_tlm_adjoint__alias__dict__", obj.__dict__)
-
-    def __new__(cls, obj):
-        obj_cls = type(obj)
-
-        class WeakAlias(cls, obj_cls):
-            pass
-
-        WeakAlias.__name__ = f"{obj_cls.__name__:s}WeakAlias"
-        return super().__new__(WeakAlias)
-
-    def __getattr__(self, key):
-        if key not in self._tlm_adjoint__alias__dict__:
-            raise AttributeError(f"No attribute '{key:s}'")
-        return self._tlm_adjoint__alias__dict__[key]
-
-    def __setattr__(self, key, value):
-        self._tlm_adjoint__alias__dict__[key] = value
-        return value
-
-    def __delattr__(self, key):
-        del self._tlm_adjoint__alias__dict__[key]
-
-    def __dir__(self):
-        return list(self._tlm_adjoint__alias__dict__.keys())
-
-
 class ControlsMarker(Equation):
     def __init__(self, M):
         """
