+++ python/tlm_adjoint/timestepping.py
@@ -20,6 +20,7 @@
 
 from .backend_interface import *
 
+from .alias import Alias
 from .base_equations import AssignmentSolver, Equation
 try:
     from .equations import EquationSolver
@@ -157,39 +158,6 @@ class TimeLevels:
 
     def __len__(self):
         return len(self._levels)
-
-
-class Alias:
-    """
-    Alias of obj, holding a reference to obj.
-    """
-
-    def __init__(self, obj):
-        if isinstance(obj, Alias):
-            raise TimesteppingException("Cannot alias Alias")
-        super().__setattr__("_tlm_adjoint__alias", obj)
-
-    def __new__(cls, obj):
-        obj_cls = type(obj)
-
-        class Alias(cls, obj_cls):
-            def __new__(cls, *args, **kwargs):
-                return obj_cls(*args, **kwargs)
-
-        Alias.__name__ = f"{obj_cls.__name__:s}Alias"
-        return super().__new__(Alias)
-
-    def __getattr__(self, key):
-        return getattr(self._tlm_adjoint__alias, key)
-
-    def __setattr__(self, key, value):
-        return setattr(self._tlm_adjoint__alias, key, value)
-
-    def __delattr__(self, key):
-        delattr(self._tlm_adjoint__alias, key)
-
-    def __dir__(self):
-        return dir(self._tlm_adjoint__alias)
 
 
 class TimeFunction:
