+++ python/tlm_adjoint_fenics/__init__.py
@@ -32,6 +32,7 @@ modules = [("backend", "tlm_adjoint_fenics"),
            ("fenics_equations", "tlm_adjoint_fenics"),
            ("tlm_adjoint", "tlm_adjoint"),
            ("backend_overrides", "tlm_adjoint_fenics"),
+           ("alias", "tlm_adjoint"),
            ("binomial_checkpointing", "tlm_adjoint"),
            ("eigendecomposition", "tlm_adjoint"),
            ("functional", "tlm_adjoint"),
