+++ python/tlm_adjoint_numpy/__init__.py
@@ -26,6 +26,7 @@ modules = [("backend", "tlm_adjoint_numpy"),
            ("backend_interface", "tlm_adjoint_numpy"),
            ("base_equations", "tlm_adjoint"),
            ("numpy_equations", "tlm_adjoint_numpy"),
+           ("alias", "tlm_adjoint"),
            ("binomial_checkpointing", "tlm_adjoint"),
            ("functional", "tlm_adjoint"),
            ("hessian", "tlm_adjoint"),
